Modularized all JavaScript code and made sure everything was imported and exported correctly. Web application runs beautifully
