+++ index.html
@@ -20,6 +20,6 @@
   <div id="card" class="card-container"></div>
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
-  <script src="main.js"></script>
+  <script type="module" src="scripts/main.js"></script>
 </body>
 </html>
