+++ scripts/main.js
@@ -1,0 +1,9 @@
+import pets from "./data.js";
+import allAnimalCards from "./allCards.js";
+import addPetBtnModal from "./formModal.js";
+import eventListenerForModal from "./eventListeners.js";
+  
+  addPetBtnModal();
+  allAnimalCards(pets);
+
+  eventListenerForModal();
