+++ scripts/renderToDom.js
@@ -1,0 +1,7 @@
+  // *********  UTILITY FUNCTIONS  ********* //
+  const renderToDom = (divId, textToRender) => {
+    const selectedElement = document.querySelector(divId);
+    selectedElement.innerHTML = textToRender;
+  };
+
+  export default renderToDom;
